# strict/swagger.json
{
    "openapi": "3.0.3",
    "paths": {
...
        "/modelVoiture/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ModelVoitureController"
                ]
            }
        },
        "/modelVoiture/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ModelVoitureController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ModelVoitureController"
                ]
            }
        },
        "/modelVoiture": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ModelVoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marqueVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelName": {
                                        "type": "string"
                                    },
                                    "marqueVoiture": {
                                        "properties": {
                                            "marqueName": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ModelVoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marqueVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelName": {
                                        "type": "string"
                                    },
                                    "marqueVoiture": {
                                        "properties": {
                                            "marqueName": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/reparation/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            }
        },
        "/reparation/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            }
        },
        "/reparation": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "startDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "endDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "avancement": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "properties": {}
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "startDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "endDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "avancement": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "properties": {}
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/voiture/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}